QoL Update!
Changed the way the listen button and the Sequence buttons behave
(They no longer can't be pressed when either a strike just occured or you just finished a stage. This fixes a rare bug where the notes would start to play multiple times)

Added 5 new Harmonys to Harmony Sequencer for a total of 6.
A random set will be chosen at th start of the bomb.
This gives the module more replayability.